+++ core/python/__init__.py
@@ -14,7 +14,7 @@ except ImportError:
 	pass
 from .util import *
 
-from .docparser import *
+from .apidoc import *
 from .dataextensions import *
 from .frameextensions import *
 from .timestreamextensions import *
